+++ opencascade-hs/src/OpenCascade/BRepOffset/Mode.hs
@@ -1,0 +1,5 @@
+module OpenCascade.BRepOffset.Mode
+(Mode (..)
+) where
+
+data Mode = Skin | Pipe | RectoVerso deriving (Enum, Show)