+++ opencascade-hs/src/OpenCascade/BRepOffsetAPI/Internal/Destructors.hs
@@ -6,5 +6,6 @@ import OpenCascade.BRepOffsetAPI.Types
 import Foreign.Ptr
 
 foreign import capi unsafe "hs_BRepOffsetAPI_MakePipe.h hs_delete_BRepOffsetAPI_MakePipe" deleteMakePipe :: Ptr MakePipe -> IO ()
+foreign import capi unsafe "hs_BRepOffsetAPI_MakeOffsetShape.h hs_delete_BRepOffsetAPI_MakeOffsetShape" deleteMakeOffsetShape :: Ptr MakeOffsetShape -> IO ()